DolphinQt: Split Config > Advanced into Misc and Advanced
It has gotten too tall.
